+++ tests/dsl/test_partial.py
@@ -1,7 +1,15 @@
 # type: ignore[all]
-from pydantic import BaseModel
-
+from pydantic import BaseModel, Field
 from instructor.dsl.partial import Partial
+import pytest
+from itertools import product
+import instructor
+from openai import OpenAI, AsyncOpenAI
+
+models = ["gpt-4o"]
+modes = [
+    instructor.Mode.TOOLS,
+]
 
 
 class SampleNestedPartial(BaseModel):
@@ -66,3 +74,58 @@ def test_partial():
 
     for model in partial.model_from_chunks(['{"b": {"b": 1}}']):
         assert model.model_dump() == {"a": None, "b": {"b": 1}}
+
+
+def test_summary_extraction():
+    class Summary(BaseModel):
+        summary: str = Field(description="A detailed summary")
+
+    client = OpenAI()
+    client = instructor.from_openai(client, mode=instructor.Mode.TOOLS)
+    extraction_stream = client.chat.completions.create_partial(
+        model="gpt-4o",
+        response_model=Summary,
+        messages=[
+            {"role": "system", "content": "You summarize text"},
+            {"role": "user", "content": "Summarize: Mary had a little lamb"},
+        ],
+        stream=True,
+    )
+
+    previous_summary = None
+    updates = 0
+    for extraction in extraction_stream:
+        if previous_summary is not None:
+            assert extraction.summary.startswith(previous_summary)
+            updates += 1
+        previous_summary = extraction.summary
+
+    assert updates > 1
+
+
+@pytest.mark.asyncio
+async def test_summary_extraction_async():
+    class Summary(BaseModel):
+        summary: str = Field(description="A detailed summary")
+
+    client = AsyncOpenAI()
+    client = instructor.from_openai(client, mode=instructor.Mode.TOOLS)
+    extraction_stream = client.chat.completions.create_partial(
+        model="gpt-4o",
+        response_model=Summary,
+        messages=[
+            {"role": "system", "content": "You summarize text"},
+            {"role": "user", "content": "Summarize: Mary had a little lamb"},
+        ],
+        stream=True,
+    )
+
+    previous_summary = None
+    updates = 0
+    async for extraction in extraction_stream:
+        if previous_summary is not None:
+            assert extraction.summary.startswith(previous_summary)
+            updates += 1
+        previous_summary = extraction.summary
+
+    assert updates > 1
